Crate futures_uds

Bindings for Unix Domain Sockets and futures

This crate provides bindings between mio_uds, the mio crate for Unix Domain sockets, and futures. The APIs and bindings in this crate are very similar to the TCP and UDP bindings in the futures-mio crate. This crate is also an empty crate on Windows, as Unix Domain Sockets are Unix-specific.

Structs

UnixDatagram
An I/O object representing a Unix datagram socket.

UnixListener
A Unix socket which can accept connections from other unix sockets.

UnixStream
A structure representing a connected unix socket.