Struct futures_uds::UnixListener

pub struct UnixListener {
    // some fields omitted
}

A Unix socket which can accept connections from other unix sockets.

Methods

impl UnixListener

fn bind<P>(path: P, handle: LoopHandle) -> IoFuture<UnixListener> where P: AsRef<Path>

Creates a new UnixListener bound to the specified path.

fn local_addr(&self) -> Result<SocketAddr>

Returns the local socket address of this listener.

fn take_error(&self) -> Result<Option<Error>>

Returns the value of the SO_ERROR option.

fn incoming(self) -> IoStream<(UnixStream, SocketAddr)>

Consumes this listener, returning a stream of the sockets this listener accepts.

This method returns an implementation of the Stream trait which resolves to the sockets the are accepted on this listener.