Struct futures_uds::UnixStream

pub struct UnixStream {
    // some fields omitted
}

A structure representing a connected unix socket.

This socket can be connected directly with UnixStream::connect or accepted from a listener with UnixListener::incoming. Additionally, a pair of anonymous Unix sockets can be created with UnixStream::pair.

Methods

impl UnixStream

fn connect<P>(p: P, handle: LoopHandle) -> IoFuture<UnixStream> where P: AsRef<Path>

Connects to the socket named by path.

This function will create a new unix socket and connect to the path specified, performing associating the returned stream with the provided event loop's handle.

The returned future will resolve once the stream is successfully connected.

fn pair(handle: LoopHandle) -> IoFuture<(UnixStream, UnixStream)>

Creates an unnamed pair of connected sockets.

This function will create a pair of interconnected unix sockets for communicating back and forth between one another. Each socket will be associated with the event loop whose handle is also provided.

fn poll_read(&self) -> Poll<(), Error>

Test whether this socket is ready to be read or not.

fn poll_write(&self) -> Poll<(), Error>

Test whether this socket is writey to be written to or not.

fn local_addr(&self) -> Result<SocketAddr>

Returns the socket address of the local half of this connection.

fn peer_addr(&self) -> Result<SocketAddr>

Returns the socket address of the remote half of this connection.

fn take_error(&self) -> Result<Option<Error>>

Returns the value of the SO_ERROR option.

fn shutdown(&self, how: Shutdown) -> Result<()>

Shuts down the read, write, or both halves of this connection.

This function will cause all pending and future I/O calls on the specified portions to immediately return with an appropriate value (see the documentation of Shutdown).